% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{abtest.uplift}
\alias{abtest.uplift}
\title{If you have a table like:}
\usage{
abtest.uplift(summary_dt, column)
}
\description{
variant   | cr
TREATMENT | 5.8
CONTROL   | 3.7
}
\details{
This function can calculate the uplift in TREATMENT using abtest.uplift(dat, 'cr').
}
