% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{datatable.with.csv}
\alias{datatable.with.csv}
\title{Datatable that hides the rownames and adds Download CSV button}
\usage{
datatable.with.csv(DT, rownames = FALSE, ...)
}
\description{
Datatable that hides the rownames and adds Download CSV button
}
