% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.ggplotly}
\alias{.ggplotly}
\title{Wrapper around ggplotly that hides the annoying Mode Bar options}
\usage{
.ggplotly(...)
}
\description{
Wrapper around ggplotly that hides the annoying Mode Bar options
}
