% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nuniq}
\alias{nuniq}
\title{Wrapper around length(unique()). The naming is influenced by nunique in pandas and in plyr, but shortend to nuniq like the UNIX tool.}
\usage{
nuniq(...)
}
\description{
Wrapper around length(unique()). The naming is influenced by nunique in pandas and in plyr, but shortend to nuniq like the UNIX tool.
}
