% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tiny.datatable}
\alias{tiny.datatable}
\title{Useful default settings for very small tabular output.}
\usage{
tiny.datatable(
  DT,
  rownames = FALSE,
  options = list(searching = FALSE, info = FALSE, paging = FALSE, ordering = FALSE)
)
}
\description{
Useful default settings for very small tabular output.
}
